workone/finetune_qwen/requirements.txt
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0
pyyaml>=6.0
sentencepiece>=0.1.99
tiktoken>=0.6.0